Cube Spawner
Added a button for spawning cubes. I intend to have buttons to spawn all other object types too. I'll need some way to spawn in objects too, which I might do by having a list that the user can add their models to, and then just type the name of the object they want to spawn or something. Maybe even give some indication of if that object is there. It's not intuitive, but it's something.

As for moving the items around the scene, I say we have one camera for UI, and one camera for looking around the scene and moving objects. The camera will likely move with the keyboard keys or something, while if selected the object is able to be moved using WASD + AE (I don't know if that's possible though).

I still need a player object, which I'll probably just setup a script for and have the user switch that script out. I'd need to make sure only one player object is allowed as well. This might conflict with moving the selected object, but I'll cross that bridge when I get to it. I also need to think about how to save the user makes.