Ending Progress
Finished the parts I was doing for the submission. I wasn't able to get the DLL working, nor was I able to track object creations or deletions for the undo/redo system. I'll address that in the video.

Now all I need to do is record the video, where I'll go over what we got done.